A flight build carries no simulation code
The host gets one cargo feature, sim, on by default; a flight build
(--no-default-features) drops the Gazebo and reference adapters, the
sidecar video client, the FC adapters camera paths, and the Aviate XIL
truth-oracle bindings. Frame identity (RawVideoFrame, UnstampedFrame,
FrameStamper) moves to pilotage-adapter-api first: the media plane
serves any producer, physical ones included, so it belongs to every
build - only the sidecar client is simulation-specific. Gated
attachments keep their Option shape through uninhabited flight-build
types, so shared code does not fork; the Aviate simulation profiles
become structurally absent (typed startup error) rather than degraded.
Clients do not change: discovery, not build variants (ADR-0026), and
the wire protocol is identical in both host builds.

Guardrail in the same change: scripts/check-flight-build.sh builds the
flight host and fails on any simulation-only crate in its dependency
tree; CI runs it. Records ADR-0040 (index row joins the ADR README
when the surface-model lane lands, which owns that file today).
