+++ hosts/session-host/src/runtime/media/transport_tests.rs
@@ -5,13 +5,13 @@
 use std::sync::Arc;
 use std::time::Duration;
 
+use pilotage_adapter_api::RawVideoFrame;
 use pilotage_adapter_api::{
     CalibrationId, CameraId, CaptureClockMapping, MeasurementClock, MeasurementStamp,
     SourceIncarnation, SourceIntegrity, SourceRole, VideoCaptureStamp,
 };
 use pilotage_protocol::wire;
 use pilotage_session::ClientKey;
-use pilotage_sim_video::RawVideoFrame;
 use pilotage_timing::SimTick;
 use tokio::sync::{mpsc, watch};
 use tokio::time::{Instant, MissedTickBehavior, timeout};
